move generate_pdf_letter func to utils app
